[docs] Fix building the API reference.
Why? docs are great

I noticed the docs have an empty API reference page; a quick check locally
seems to confirm the problem is that the usual sphinx `sys.path` adjustments
predate moving the source into `src/hypercorn`; this PR fixes the path
accordingly.

(Happy to _remove_ the `../` entry too, but wasn't sure if something else has
quietly grown a dep on that in the docs.)